Rename `DocumentListSeparator` to `AdjacentListsSupport` and add documentation.
