+++ packages/ckeditor5-list/src/index.ts
@@ -10,7 +10,7 @@
 export { default as DocumentList } from './documentlist';
 export { default as DocumentListEditing, type DocumentListEditingPostFixerEvent } from './documentlist/documentlistediting';
 export { default as DocumentListIndentCommand } from './documentlist/documentlistindentcommand';
-export { default as DocumentListSeparator } from './documentlist/documentlistseparator';
+export { default as AdjacentListsSupport } from './documentlist/adjacentlistssupport';
 export { default as DocumentListProperties } from './documentlistproperties';
 export { default as DocumentListPropertiesEditing } from './documentlistproperties/documentlistpropertiesediting';
 export { default as DocumentListUtils } from './documentlist/documentlistutils';
